Separate Slang runtime from REPL implementation. This allows useres to write their own languages targeting a generic sabre REPL implementation.  See cmd/slang/main.go for an example.
