+++ go.mod
@@ -3,10 +3,10 @@ module github.com/spy16/sabre
 go 1.13
 
 require (
-	github.com/SentimensRG/ctx v0.0.0-20180729130232-0bfd988c655d
+	github.com/chzyer/logex v1.1.10 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
+	github.com/chzyer/test v0.0.0-20180213035817-a1ea475d72b1 // indirect
 	github.com/lthibault/log v0.0.0-20190513014217-f549b3a28a20
 	github.com/sirupsen/logrus v1.4.2 // indirect
-	github.com/spy16/parens v1.0.0
 	golang.org/x/sys v0.0.0-20200117145432-59e60aa80a0c // indirect
 )
